pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "defiguard"
version = "0.1.0"
description = "DeFiGuard CCS artifact bundle (minimal reproduction + toy implementation)"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "pandas>=2.0",
  "numpy>=1.24",
  "pyyaml>=6.0",
  "matplotlib>=3.7",
  "tqdm>=4.66",
]

[project.scripts]
defiguard = "defiguard.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
